PRJ-20 Add a hacky workaround that hides the scrolling bug in many cases
Seems like the bug happens because we don't have any synchronisation between rendering to different surfaces. As for scroll panes an offscreen surface is used and it's firstly updated and then rendered to an on-screen surface, synchronization really matters. This hack contains heuristics of sorting drawing to surfaces in the correct order, it practically makes the situation much better but it can be false sometimes so the bug isn't fixed. Also, the hack won't work if there are too many drawings of one surface to another in a short period of time – we just have a buffer of about 10 ms on the server. We need to implement drawing order synchronization between different surfaces to fully fix the issue