Add support for disabling font ligatures
Issue:
fonts are rendered with ligatures always enabled, even when users disable the
ligatures from Settings -> Editor -> Font

Fix:
Font events will now contain a flag which specifies if ligatures are disabled or not.
This flag will be used by the DOM renderer to set the font-variant-ligatures
property to `none` or `normal` for the canvas.

For now there is an issue with font changing x-height when switching the ligatures on or off.
Initially I thought the problem is because glyphs are rendered using the path fill primitives
from DOM while simple text is using the native text drawing primitives. However, after I played
with the settings the issue reproduces ONLY with JetBrains Mono family and ONLY at sizes of 13px
and below. There is no issue if I increase the size or change the font.

I think the problem lies in the font itself, though weirdly enough I can't reproduce it natively with
latest IntelliJ. Maybe we need to upgrade the mono fonts provided by projector-server to latest version.