Implement basic window draw interest management for server
While this feature is of little use for the web client currently, it is needed in IDEA/CWM to reduce network traffic overhead when the main IDEA window is not visible on the client

This implementation assumes that the client is interested in all windows by default, and lets the client mark specific windows as non-interesting. Draw events for non-interesting windows will be discarded when sent to that client.
Marking the window as interesting again will trigger a repaint to ensure that the client has up-to-date window state.